+++ Controller/ReviewerController.cs
@@ -11,10 +11,14 @@ namespace Pokemon.Controller
     public class ReviewerController : ControllerBase
     {
         private readonly IReviewerRepository _reviewerRepository;
+        private readonly IReviewRepository _reviewRepository;
         private readonly IMapper _mapper;
-        public ReviewerController(IReviewerRepository reviewerRepository, IMapper mapper)
+        public ReviewerController(IReviewerRepository reviewerRepository, 
+        IReviewRepository reviewRepository,
+        IMapper mapper)
         {
             _reviewerRepository = reviewerRepository;
+            _reviewRepository = reviewRepository;
             _mapper = mapper;
         }
 
@@ -122,8 +126,16 @@ namespace Pokemon.Controller
                 return NotFound();
 
             var reviewerToDelete = _reviewerRepository.GetReviewer(reviewerId);
+            var rewiewsToDelete = _reviewerRepository.GetReviewsByReviewer(reviewerId);
+
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            if (!_reviewRepository.DeleteReviews(rewiewsToDelete.ToList()))
+            {
+                ModelState.AddModelError("", "Something went wrong while deleting reviews");
+                return StatusCode(500, ModelState);
+            }
 
             if (!_reviewerRepository.DeleteReviewer(reviewerToDelete))
             {
